Perf w/inserts. Check binds & use schema cache. Fixes #572. Thanks @noelr.
