+++ lib/active_record/connection_adapters/sqlserver/schema_statements.rb
@@ -497,27 +497,6 @@ module ActiveRecord
           match_data ? match_data[1] : column_name
         end
 
-        # === SQLServer Specific (Identity Inserts) ===================== #
-
-        def query_requires_identity_insert?(sql)
-          if insert_sql?(sql)
-            table_name = get_table_name(sql)
-            id_column = identity_columns(table_name).first
-            id_column && sql =~ /^\s*(INSERT|EXEC sp_executesql N'INSERT)[^(]+\([^)]*\b(#{id_column.name})\b,?[^)]*\)/i ? quote_table_name(table_name) : false
-          else
-            false
-          end
-        end
-
-        def insert_sql?(sql)
-          !(sql =~ /^\s*(INSERT|EXEC sp_executesql N'INSERT)/i).nil?
-        end
-
-        def identity_columns(table_name)
-          columns(table_name).select(&:is_identity?)
-        end
-
-
         private
 
         def create_table_definition(*args)
