+++ test/cases/scratchpad_test_sqlserver.rb
@@ -1,9 +1,11 @@
 require 'cases/helper_sqlserver'
+require 'models/book'
 
 class ScratchpadTestSQLServer < ActiveRecord::TestCase
 
   it 'helps debug things' do
-    #
+    $FOO = true
+    1000.times { Book.create! name: 'test' }
   end
 
 end
